fix: tables updated.

+++ db/migration/000004_create_table_reply.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE "Reply" (
-                         id SERIAL PRIMARY KEY,
-                         user_id INT REFERENCES "User"(id) ON DELETE CASCADE,
-                         comment_id INT REFERENCES "Comment"(id) ON DELETE CASCADE,
+                         id BIGSERIAL PRIMARY KEY,
+                         user_id BIGINT REFERENCES "User"(id) ON DELETE CASCADE,
+                         comment_id BIGINT REFERENCES "Comment"(id) ON DELETE CASCADE,
                          content VARCHAR NOT NULL,
                          likes INT DEFAULT 0,
                          created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
