+++ db/migration/000006_create_table_follower.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE "Follower" (
-                            id SERIAL PRIMARY KEY,
-                            follower_id INT REFERENCES "User"(id) ON DELETE CASCADE,
-                            followed_id INT REFERENCES "User"(id) ON DELETE CASCADE,
+                            id BIGSERIAL PRIMARY KEY,
+                            follower_id BIGINT REFERENCES "User"(id) ON DELETE CASCADE,
+                            followed_id BIGINT REFERENCES "User"(id) ON DELETE CASCADE,
                             created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
